requirements.txt
accelerate==0.34.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blis==0.7.11
bson==0.5.10
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
cryptography==43.0.0
cymem==2.0.8
dataclasses-json==0.6.7
datasets==2.21.0
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
einops==0.8.0
email_validator==2.2.0
evaluate==0.4.2
Events==0.5
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.112.1
fastapi-jwt-auth==0.5.0
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.7
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.39
langchain-huggingface==0.0.3
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.117
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
motor==3.5.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.19.0
openai==1.45.1
opencv-python==4.10.0.84
opensearch-py==2.6.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==1.7.1
pymongo==4.8.0
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
pytest==8.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
rapidocr-onnxruntime==1.3.24
regex==2024.7.24
requests==2.32.3
rich==13.8.0
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
seqeval==1.2.2
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.2
sympy==1.13.2
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
triton==3.0.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.4
